Use SPS parsing also for getting information out of MP4
